Only report libvirt_domain_interface_meta if exist
This commit changes the way that we report
`libvirt_domain_interface_meta` metric so that it is only present if the
`SourceBridge` or `VirtualInterface` are present. In systems which use
SR-IOV, there is no concept of a bridge or virtual interface, as the
interface is completely passed through to the hypervisor.

This resolves a problem where the exporter would fail on a system with
SR-IOV with the following error:

```
An error has occurred while serving metrics:

collected metric "libvirt_domain_interface_meta" { label:<name:"domain" value:"host" > label:<name:"source_bridge" value:"" > label:<name:"target_device" value:"" > label:<name:"virtual_interface" value:"" > gauge:<value:1 > } was collected before with the same name and label values
```